Inline TriangleKdTree into remeshing implementation
